Broke Reddit into its own view
